acc2: Move AcceleratorInfo to new python module
